V103.00	Adopted FatFs0.15 latest patch update(patch3). 		Adopted latest FONTX2 Drivers. 		Adopted CMSIS Core(M) Version 6 Header Files into latest repository. 		 -> https://github.com/ARM-software/CMSIS_6/commit/f09017100a0eca8ccd5769c794d2de77f511c823 		Adopted arm-gcc toolchain updates 		 -> (Arm GNU Toolchain 12.3.Rel1 (Build arm-12.35)) 12.3.1 20230626

+++ lib/CMSIS/Core/Include/core_cm23.h
@@ -1826,7 +1826,7 @@ __STATIC_INLINE void NVIC_DecodePriority (uint32_t Priority, uint32_t PriorityGr
 __STATIC_INLINE void __NVIC_SetVector(IRQn_Type IRQn, uint32_t vector)
 {
 #if defined (__VTOR_PRESENT) && (__VTOR_PRESENT == 1U)
-  uint32_t *vectors = (uint32_t *)SCB->VTOR;
+  uint32_t *vectors = (uint32_t *) ((uintptr_t) SCB->VTOR);
 #else
   uint32_t *vectors = (uint32_t *)0x0U;
 #endif
@@ -1846,7 +1846,7 @@ __STATIC_INLINE void __NVIC_SetVector(IRQn_Type IRQn, uint32_t vector)
 __STATIC_INLINE uint32_t __NVIC_GetVector(IRQn_Type IRQn)
 {
 #if defined (__VTOR_PRESENT) && (__VTOR_PRESENT == 1U)
-  uint32_t *vectors = (uint32_t *)SCB->VTOR;
+  uint32_t *vectors = (uint32_t *) ((uintptr_t) SCB->VTOR);
 #else
   uint32_t *vectors = (uint32_t *)0x0U;
 #endif
@@ -2048,7 +2048,7 @@ __STATIC_INLINE uint32_t TZ_NVIC_GetPriority_NS(IRQn_Type IRQn)
 
 #if defined (__MPU_PRESENT) && (__MPU_PRESENT == 1U)
 
-#include "mpu_armv8.h"
+  #include "m-profile/armv8m_mpu.h"
 
 #endif
